Decouple argument parsing/business logic
